### 2026-07-15 #10961 @editorialbot
[REVIEW]: pysqa - Python Simple Queuing System Adapter
**Submitting author:** @jan-janssen (<a href="http://orcid.org/0000-0001-9948-7119">Jan Janssen</a>)
**Repository:** https://github.com/pyiron/pysqa
**Branch with paper.md** (empty if default branch): main
**Version:** 1.0.0
