### 2026-07-18 d08ccfe
Update tqdm to 4.69.0 in pyproject.toml and environment files
