### 2026-07-18 1213da4
Update tqdm to 4.69.0 in pyproject.toml and environment files (#546)
Co-authored-by: copilot-swe-agent[bot] <198982749+Copilot@users.noreply.github.com>
