+++ .ci_support/environment-docs.yml
@@ -10,7 +10,7 @@ dependencies:
 - pyyaml =6.0.3
 - jinja2 =3.1.6
 - paramiko =5.0.0
-- tqdm =4.69.0
+- tqdm =4.69.1
 - jupyter-book =1.0.0
 - python =3.12
 - pydantic =2.13.4
