+++ .ci_support/environment.yml
@@ -7,7 +7,7 @@ dependencies:
 - pyyaml =6.0.3
 - jinja2 =3.1.6
 - paramiko =5.0.0
-- tqdm =4.69.0
+- tqdm =4.69.1
 - hatchling =1.31.0
 - hatch-vcs =0.5.0
 - pydantic =2.13.4
