### 2026-07-27 6fe9a8b
Bump pyauthenticator from 0.3.1 to 0.4.0 (#551)
Bumps [pyauthenticator](https://github.com/jan-janssen/pyauthenticator) from 0.3.1 to 0.4.0.
- [Release notes](https://github.com/jan-janssen/pyauthenticator/releases)
- [Commits](https://github.com/jan-janssen/pyauthenticator/compare/pyauthenticator-0.3.1...pyauthenticator-0.4.0)

---
updated-dependencies:
- dependency-name: pyauthenticator
  dependency-version: 0.4.0
  dependency-type: direct:development
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+
[…]
