+++ pyproject.toml
@@ -49,7 +49,7 @@ remote = [
     "paramiko==5.0.0",
     "tqdm==4.69.1",
 ]
-twofactor = ["pyauthenticator==0.3.1"]
+twofactor = ["pyauthenticator==0.4.0"]
 pydantic = ["pydantic==2.13.4"]
 
 [project.scripts]
