### 2026-07-27 7f0521d
Bump tqdm from 4.69.1 to 4.70.0
Bumps [tqdm](https://github.com/tqdm/tqdm) from 4.69.1 to 4.70.0.
- [Release notes](https://github.com/tqdm/tqdm/releases)
- [Commits](https://github.com/tqdm/tqdm/compare/v4.69.1...v4.70.0)

---
updated-dependencies:
- dependency-name: tqdm
  dependency-version: 4.70.0
  dependency-type: direct:development
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
