### 2026-07-28 dcee882
Add PLR0917 to linting configuration
