+++ pyproject.toml
@@ -92,6 +92,8 @@ ignore = [
     "PLR0912",
     # Too many statements
     "PLR0915",
+    # Too many not-keyword-only arguments,
+    "PLR0917",
 ]
 
 [tool.hatch.build]
