### 2026-07-29 de63bff
Update pandas to 3.0.5 (#548)
