+++ pyproject.toml
@@ -1,7 +1,7 @@
 [build-system]
 requires = [
     "jinja2==3.1.6", 
-    "pandas==3.0.3", 
+    "pandas==3.0.5", 
     "pyyaml==6.0.3", 
     "hatchling==1.31.0",
     "hatch-vcs==0.5.0",
@@ -33,7 +33,7 @@ classifiers = [
 ]
 dependencies = [
     "jinja2==3.1.6",
-    "pandas==3.0.3",
+    "pandas==3.0.5",
     "pyyaml==6.0.3",
 ]
 dynamic = ["version"]
