+++ .ci_support/environment.yml
@@ -3,7 +3,7 @@ channels:
 dependencies:
 - defusedxml =0.7.1
 - coverage
-- pandas =3.0.3
+- pandas =3.0.5
 - pyyaml =6.0.3
 - jinja2 =3.1.6
 - paramiko =5.0.0
