+++ .ci_support/environment-mini.yml
@@ -1,7 +1,7 @@
 channels:
 - conda-forge
 dependencies:
-- pandas =3.0.3
+- pandas =3.0.5
 - pyyaml =6.0.3
 - jinja2 =3.1.6
 - hatchling =1.31.0
