+++ .ci_support/environment-docs.yml
@@ -6,7 +6,7 @@ dependencies:
 - sphinx_rtd_theme
 - myst-parser
 - defusedxml =0.7.1
-- pandas =3.0.3
+- pandas =3.0.5
 - pyyaml =6.0.3
 - jinja2 =3.1.6
 - paramiko =5.0.0
