+++ .ci_support/environment-docs.yml
@@ -14,5 +14,5 @@ dependencies:
 - jupyter-book =1.0.0
 - python =3.12
 - pydantic =2.13.4
-- hatchling =1.31.0
+- hatchling =1.32.0
 - hatch-vcs =0.5.0
