### 2026-08-14 a066b0f
Bump pyauthenticator from 0.4.0 to 0.4.1 (#557)
Bumps [pyauthenticator](https://github.com/jan-janssen/pyauthenticator) from 0.4.0 to 0.4.1.
- [Release notes](https://github.com/jan-janssen/pyauthenticator/releases)
- [Commits](https://github.com/jan-janssen/pyauthenticator/compare/pyauthenticator-0.4.0...pyauthenticator-0.4.1)

---
updated-dependencies:
- dependency-name: pyauthenticator
  dependency-version: 0.4.1
  dependency-type: direct:development
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+
[…]
