+++ pyproject.toml
@@ -49,7 +49,7 @@ remote = [
     "paramiko==5.0.0",
     "tqdm==4.70.0",
 ]
-twofactor = ["pyauthenticator==0.4.0"]
+twofactor = ["pyauthenticator==0.4.1"]
 pydantic = ["pydantic==2.13.4"]
 
 [project.scripts]
