### 2026-08-14 e2688a4
Bump hatchling from 1.31.0 to 1.32.0 (#558)
* Bump hatchling from 1.31.0 to 1.32.0

Bumps [hatchling](https://github.com/pypa/hatch) from 1.31.0 to 1.32.0.
- [Release notes](https://github.com/pypa/hatch/releases)
- [Commits](https://github.com/pypa/hatch/compare/hatchling-v1.31.0...hatchling-v1.32.0)

---
updated-dependencies:
- dependency-name: hatchling
  dependency-version: 1.32.0
  dependency-type: direct:development
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>

* [dependabot skip] Update environment

---------

Signed-off-by: dependa
[…]
