+++ binder/environment.yml
@@ -9,6 +9,6 @@ dependencies:
 - jinja2 =3.1.6
 - paramiko =5.0.0
 - tqdm =4.70.0
-- hatchling =1.31.0
+- hatchling =1.32.0
 - hatch-vcs =0.5.0
 - pydantic =2.13.4
