+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = [
     "jinja2==3.1.6", 
     "pandas==3.0.5", 
     "pyyaml==6.0.3", 
-    "hatchling==1.31.0",
+    "hatchling==1.32.0",
     "hatch-vcs==0.5.0",
 ]
 build-backend = "hatchling.build"
