### 2026-08-23 67f2519
Fix spelling mistakes in JOSS draft
