### 2026-08-23 8a6d2ad
Update URL for Adaptive Computing entry in bibliography (#561)
