+++ docs/paper/paper.bib
@@ -141,7 +141,7 @@
 
 @misc{moab, 
     title = {Adaptive Computing},
-    url={https://adaptivecomputing.com/cherry-services/torque-resource-manager/},
+    url={https://adaptivecomputing.com/moab-hpc-suite},
     year={2003}
 }
 
