### 2026-08-23 2b4d640
Clarify target audience and benefits of pysqa
Expanded the statement of need to clarify the target audience for pysqa and its advantages over embedding scheduler-specific commands.
