+++ docs/paper/paper.md
@@ -29,7 +29,7 @@ pysqa (Python Simple Queue Adapter) is an open-source Python library that provid
 Unlike workflow orchestration frameworks, pysqa focuses exclusively on the scheduler interaction layer. It enables computational scientists and research software engineers to integrate HPC resources into Python applications with minimal complexity while retaining full control over scheduler-specific configurations. A schematic overview of pysqa is illustrated in \autoref{fig:pysqa}.
 
 # Statement of Need
-Modern computational research increasingly relies on automated execution of simulations, machine learning workloads, and data-processing pipelines on shared HPC infrastructure. While scheduler command-line tools such as sbatch, squeue, and scancel provide direct access to HPC resources, scientific applications often require programmatic job submission and monitoring capabilities. Embedding scheduler-specific commands directly into software reduces portability and increases maintenance costs when users operate across multiple clusters.
+Modern computational research increasingly relies on automated execution of simulations, machine learning workloads, and data-processing pipelines on shared HPC infrastructure. While scheduler command-line tools such as sbatch, squeue, and scancel provide direct access to HPC resources, scientific applications often require programmatic job submission and monitoring capabilities. Embedding scheduler-specific commands directly into software reduces portability and increases maintenance costs when users operate across multiple clusters. The primary audience for pysqa is developers of scientific software and research software engineers who require a lightweight, programmatic interface to HPC schedulers that can be used across different computing environments.
 
 # State of the field 
 Several software projects address related challenges. MyQueue [@myqueue] provides a higher-level task and workflow abstraction designed for scientific computing campaigns. PSI/J [@psij] offers a portable job execution API spanning multiple schedulers and execution backends. Jobflow-Remote focuses on remote execution of workflow graphs within the Jobflow ecosystem [@jobflow]. These tools provide broader workflow or interoperability capabilities, but they also introduce additional abstractions and infrastructure requirements.
