+++ tests/integration/test_remote.py
@@ -26,9 +26,9 @@ class TestRemoteQueueAdapter(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         cls.path = os.path.dirname(os.path.abspath(__file__))
-        cls.remote = QueueAdapter(directory=os.path.join(cls.path, "../../static/remote"))
+        cls.remote = QueueAdapter(directory=os.path.join(cls.path, "../static/remote"))
         cls.remote_alternative = QueueAdapter(
-            directory=os.path.join(cls.path, "../../static/remote_alternative")
+            directory=os.path.join(cls.path, "../static/remote_alternative")
         )
 
     def test_config(self):
@@ -136,7 +136,7 @@ class TestRemoteQueueAdapter(unittest.TestCase):
 class TestRemoteQueueAdapterRebex(unittest.TestCase):
     def test_remote_command_individual_connections(self):
         path = os.path.dirname(os.path.abspath(__file__))
-        remote = QueueAdapter(directory=os.path.join(path, "../../static/remote_rebex"))
+        remote = QueueAdapter(directory=os.path.join(path, "../static/remote_rebex"))
         remote._adapter._ssh_remote_path = path
         remote._adapter._open_ssh_connection()
         output = remote._adapter._execute_remote_command(command="pwd")
@@ -144,7 +144,7 @@ class TestRemoteQueueAdapterRebex(unittest.TestCase):
 
     def test_remote_command_continous_connection(self):
         path = os.path.dirname(os.path.abspath(__file__))
-        remote = QueueAdapter(directory=os.path.join(path, "../../static/remote_rebex"))
+        remote = QueueAdapter(directory=os.path.join(path, "../static/remote_rebex"))
         remote._adapter._ssh_remote_path = path
         remote._adapter._ssh_continous_connection = True
         remote._adapter._open_ssh_connection()
@@ -153,7 +153,7 @@ class TestRemoteQueueAdapterRebex(unittest.TestCase):
 
     def test_remote_command_individual_connections_hosts(self):
         path = os.path.dirname(os.path.abspath(__file__))
-        remote = QueueAdapter(directory=os.path.join(path, "../../static/remote_rebex_hosts"))
+        remote = QueueAdapter(directory=os.path.join(path, "../static/remote_rebex_hosts"))
         remote._adapter._ssh_remote_path = path
         remote._adapter._open_ssh_connection()
         output = remote._adapter._execute_remote_command(command="pwd")
@@ -161,7 +161,7 @@ class TestRemoteQueueAdapterRebex(unittest.TestCase):
 
     def test_remote_command_continous_connection_hosts(self):
         path = os.path.dirname(os.path.abspath(__file__))
-        remote = QueueAdapter(directory=os.path.join(path, "../../static/remote_rebex_hosts"))
+        remote = QueueAdapter(directory=os.path.join(path, "../static/remote_rebex_hosts"))
         remote._adapter._ssh_remote_path = path
         remote._adapter._ssh_continous_connection = True
         remote._adapter._open_ssh_connection()
@@ -170,34 +170,34 @@ class TestRemoteQueueAdapterRebex(unittest.TestCase):
 
     def test_submit_job(self):
         path = os.path.dirname(os.path.abspath(__file__))
-        remote = QueueAdapter(directory=os.path.join(path, "../../static/remote_rebex_hosts"))
+        remote = QueueAdapter(directory=os.path.join(path, "../static/remote_rebex_hosts"))
         remote._adapter._ssh_remote_path = path
-        output = remote._adapter.submit_job(working_directory=os.path.join(path, "../../static/empty"), command="echo 1")
+        output = remote._adapter.submit_job(working_directory=os.path.join(path, "../static/empty"), command="echo 1")
[…]
