### 2026-08-23 17bf1d2
Move remote tests to integration tests (#563)
* Move remote tests to integration tests

* fix path
