+++ tests/unit/base/test_remote_ssh_connection.py
@@ -1,489 +1,0 @@
-import os
-import unittest
-from unittest.mock import MagicMock, patch
-
-from pysqa import QueueAdapter
-
-try:
-    import paramiko
-    from tqdm import tqdm
-    from pysqa.base.remote import get_transport
-
-    skip_remote_test = False
-except ImportError:
-    skip_remote_test = True
-
-
-class FakeSSH:
-    @staticmethod
-    def get_transport(*args, **kwargs):
-        return None
-
-
-def _new_remote(directory_name):
-    path = os.path.dirname(os.path.abspath(__file__))
-    return QueueAdapter(directory=os.path.join(path, "../../static", directory_name))
-
-
-@unittest.skipIf(
-    skip_remote_test,
-    "Either paramiko or tqdm are not installed, so the remote queue adapter tests are skipped.",
-)
-class TestGetTransport(unittest.TestCase):
-    """Pure unit tests for get_transport(), without any network access."""
-
-    def test_get_transport_returns_transport(self):
-        ssh = MagicMock()
-        ssh.get_transport.return_value = "a-transport"
-        self.assertEqual(get_transport(ssh=ssh), "a-transport")
-
-    def test_get_transport_raises_value_error_when_missing(self):
-        with self.assertRaises(ValueError):
-            get_transport(ssh=FakeSSH())
-
-
-@unittest.skipIf(
-    skip_remote_test,
-    "Either paramiko or tqdm are not installed, so the remote queue adapter tests are skipped.",
-)
-class TestOpenSSHConnection(unittest.TestCase):
-    """Unit tests for the authentication branches of _open_ssh_connection().
-
-    paramiko.SSHClient is mocked out entirely, so these tests never touch the
-    network - unlike the TestRemoteQueueAdapterRebex tests which connect to
-    test.rebex.net and therefore require network access plus a known_hosts
-    entry that matches the CI setup.
-    """
-
-    def test_key_with_passphrase(self):
-        remote = _new_remote("remote")
-        remote._adapter._ssh_key_passphrase = "secret"
-        with patch("pysqa.base.remote.paramiko.SSHClient") as mock_cls:
-            mock_ssh = MagicMock()
-            mock_cls.return_value = mock_ssh
-            result = remote._adapter._open_ssh_connection()
-        self.assertIs(result, mock_ssh)
-        mock_ssh.connect.assert_called_once_with(
-            hostname="hpc-cluster.university.edu",
-            port=22,
-            username="hpcuser",
-            key_filename=remote._adapter._ssh_key,
-            passphrase="secret",
-        )
-
-    def test_key_without_passphrase(self):
-        remote = _new_remote("remote")
-        with patch("pysqa.base.remote.paramiko.SSHClient") as mock_cls:
-            mock_ssh = MagicMock()
-            mock_cls.return_value = mock_ssh
-            result = remote._adapter._open_ssh_connection()
-        self.assertIs(result, mock_ssh)
-        mock_ssh.connect.assert_called_once_with(
-            hostname="hpc-cluster.university.edu",
-            port=22,
-            username="hpcuser",
-            key_filename=remote._adapter._ssh_key,
-        )
-
-    def test_password_auth(self):
-        remote = _new_remote("remote")
-        remote._adapter._ssh_key = None
-        remote._adapter._ssh_password = "test1234"
-        with patch("pysqa.base.remote.paramiko.SSHClient") as mock_cls:
-            mock_ssh = MagicMock()
-            mock_cls.return_value = mock_ssh
-            result = remote._adapter._open_ssh_connection()
-        self.assertIs(result, mock_ssh)
-        mock_ssh.connect.assert_called_once_with(
-            hostname="hpc-cluster.university.edu",
[…]
