+++ tests/unit/base/test_remote.py
@@ -1,5 +1,7 @@
 import os
 import unittest
+from unittest.mock import MagicMock, patch
+
 from pysqa import QueueAdapter
 
 try:
@@ -18,198 +20,470 @@ class FakeSSH:
         return None
 
 
-@unittest.skipIf(
-    skip_remote_test,
-    "Either paramiko or tqdm are not installed, so the remote queue adapter tests are skipped.",
-)
-class TestRemoteQueueAdapter(unittest.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        cls.path = os.path.dirname(os.path.abspath(__file__))
-        cls.remote = QueueAdapter(directory=os.path.join(cls.path, "../../static/remote"))
-        cls.remote_alternative = QueueAdapter(
-            directory=os.path.join(cls.path, "../../static/remote_alternative")
-        )
-
-    def test_config(self):
-        self.assertEqual(self.remote.config["queue_type"], "REMOTE")
-        self.assertEqual(self.remote.config["queue_primary"], "remote")
-
-    def test_list_clusters(self):
-        self.assertEqual(self.remote.list_clusters(), ["default"])
-
-    def test_remote_flag(self):
-        self.assertTrue(self.remote._adapter.remote_flag)
-        self.assertTrue(self.remote_alternative._adapter.remote_flag)
-
-    def test_create_remote_dir(self):
-        with self.assertRaises(TypeError):
-            self.remote._adapter._create_remote_dir(directory=None)
-
-    def test_ssh_delete_file_on_remote(self):
-        self.assertEqual(self.remote.ssh_delete_file_on_remote, False)
-
-    def test_ssh_continous_connection(self):
-        self.assertEqual(self.remote._adapter._ssh_continous_connection, True)
-        self.assertEqual(
-            self.remote_alternative._adapter._ssh_continous_connection, False
-        )
-
-    def test_submit_job_remote(self):
-        with self.assertRaises(NotImplementedError):
-            self.remote.submit_job(queue="remote", dependency_list=[])
-
-    def test_submit_command(self):
-        with self.subTest("remote config"):
-            command = self.remote._adapter._submit_command(
-                queue="remote",
-                job_name="test",
-                working_directory="/home/localuser/projects/test",
-                cores=str(1),
-                memory_max=str(1),
-                run_time_max=str(1),
-                command_str="/bin/true",
-            )
-            self.assertEqual(
-                command,
-                'python -m pysqa --config_directory /u/share/pysqa/resources/queues/ --submit --queue remote --job_name test --working_directory /home/localuser/projects/test --cores 1 --memory 1 --run_time 1 --command "/bin/true" ',
-            )
-        with self.subTest("alternative remote config"):
-            command = self.remote_alternative._adapter._submit_command(
-                queue="remote",
-                job_name="test",
-                working_directory="/home/localuser/projects/test",
-                cores=str(1),
-                memory_max=str(1),
-                run_time_max=str(1),
-                command_str="/bin/true",
-            )
-            self.assertEqual(
-                command,
-                'python3 -m pysqa --config_directory /u/share/pysqa/resources/queues/ --submit --queue remote --job_name test --working_directory /home/localuser/projects/test --cores 1 --memory 1 --run_time 1 --command "/bin/true" ',
-            )
-
-    def test_get_queue_status_command(self):
-        command = self.remote._adapter._get_queue_status_command()
-        self.assertEqual(
-            command,
-            "python -m pysqa --config_directory /u/share/pysqa/resources/queues/ --status",
-        )
-
-    def test_convert_path_to_remote(self):
[…]
