### 2026-08-24 #10961 @AnjaliSandip
@GuilloteauQ @CompEng0001 @jmukher1 Is the paper/software ready for publication?
