### 2026-08-24 25f49b6
[pre-commit.ci] pre-commit autoupdate (#564)
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.16.3 → v0.16.4](https://github.com/astral-sh/ruff-pre-commit/compare/v0.16.3...v0.16.4)

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
