### 2026-08-25 #10961 @CompEng0001
@jan-janssen FYI there is also a formatting error in `README.md`:

- https://github.com/pyiron/pysqa/blob/25f49b610a46920df934c22567c784922a1eb531/README.md?plain=1#L56

  ```md
    * [LSF]](https://pysqa.readthedocs.io/en/latest/queue.html#lsf)
  ```
    * [LSF]](https://pysqa.readthedocs.io/en/latest/queue.html#lsf)
