+++ docs/paper/paper.bib
@@ -150,3 +150,9 @@
     url={https://www.openpbs.org},
     year={1998}
 }
+
+@misc{jinja2, 
+    title = {Jinja},
+    url={https://jinja.palletsprojects.com},
+    year={2008}
+}
