+++ docs/paper/paper.md
@@ -24,7 +24,7 @@ High-performance computing (HPC) resources are commonly accessed through batch s
 
 ![Schematic overview of pysqa. The package provides a scheduler-independent interface for submitting, monitoring, and managing HPC jobs from Python or the command line. Reusable Jinja2-based templates are translated to the native syntax of supported queuing systems (e.g., SLURM, PBS, LSF, and SGE), while built-in support for remote connections, reservations, and job management simplifies the integration of HPC resources into scientific workflows.\label{fig:pysqa}](pysqa.png){width="100%"}
 
-pysqa (Python Simple Queue Adapter) is an open-source Python library that provides a lightweight, scheduler-independent interface for submitting, monitoring, and managing HPC jobs. The design of pysqa is based on reusable queue templates written in the Jinja2 templating language. Users define scheduler-specific submission scripts once and subsequently reuse them across many computational tasks and workflows. This approach separates scheduler configuration from application logic while preserving transparency by generating native scheduler submission scripts. pysqa currently supports Flux, LSF, Moab, SGE, SLURM, and Torque queuing systems and includes optional support for remote job submission via SSH.
+pysqa (Python Simple Queue Adapter) is an open-source Python library that provides a lightweight, scheduler-independent interface for submitting, monitoring, and managing HPC jobs. The design of pysqa is based on reusable queue templates written in the Jinja2 templating language [@jinja2]. Users define scheduler-specific submission scripts once and subsequently reuse them across many computational tasks and workflows. This approach separates scheduler configuration from application logic while preserving transparency by generating native scheduler submission scripts. pysqa currently supports Flux, LSF, Moab, SGE, SLURM, and Torque queuing systems and includes optional support for remote job submission via SSH.
 
 Unlike workflow orchestration frameworks, pysqa focuses exclusively on the scheduler interaction layer. It enables computational scientists and research software engineers to integrate HPC resources into Python applications with minimal complexity while retaining full control over scheduler-specific configurations. A schematic overview of pysqa is illustrated in \autoref{fig:pysqa}.
 
