### 2026-08-25 c1f5e84
Fix formatting error in the readme (#565)
