### 2026-08-25 #10961 @jan-janssen
@CompEng0001 Thanks for the feedback, I fixed the `README.md` issue in https://github.com/pyiron/pysqa/pull/565 and added the citation for `jinja2`.
