### 2026-08-25 #10961 @jmukher1
There seems to be some editing (may be Copy-paste) errors in docs/queue.md. I would like the authors to confirm it as either intentional or just fix them appropriately.

Every section from LSF onward closes with the same sentence saying "the flux interface can be enabled" — LSF, MOAB, SGE, SLURM and TORQUE all say "flux". Should not they should name their own system?

The SLURM section tells you to import from pysqa.wrapper.sge import template and set queue_type="sge". Seems like a Wrong module and wrong value. Please cross-check.

The TORQUE section does the same thing one step further off: from pysqa.wrapper.slurm import template, queue_type="slurm".  Please cross-check.

Flux points at pysqa.wrapper.flux, but the API sidebar lists only abstract, gent, lsf, moab, sge, slurm, torque under pysqa.wrapper — no flux module. Conversely pysqa.wrapper.gent exists in the API but is documented nowhere in the queue page or the README's supported-systems list.
The queue.yaml skeleton enumerates FLUX, LSF, MOAB, SGE, SLURM, and drops TORQUE, which the README advertises as supported.

Minor: Docs footer still reads © 2023. Kindly take a look.
