### 2026-08-25 81985a5
Add copyright year to configuration file (#567)
