+++ docs/_config.yml
@@ -1,5 +1,6 @@
 title: pysqa
 author: Jan Janssen
+copyright: "2026"
 
 execute:
   execute_notebooks           : off
