### 2026-08-25 #10961 @jan-janssen
@jmukher1 thanks for your feedback, I updated the documentation accordingly:

> There seems to be some editing (may be Copy-paste) errors in docs/queue.md. I would like the authors to confirm it as either intentional or just fix them appropriately.

This was addressed in https://github.com/pyiron/pysqa/pull/566

> Minor: Docs footer still reads © 2023. Kindly take a look.

This was addressed in https://github.com/pyiron/pysqa/pull/567
