### 2026-08-31 7fc1da0
Bump pydantic from 2.13.4 to 2.13.5 (#568)
* Bump pydantic from 2.13.4 to 2.13.5

Bumps [pydantic](https://github.com/pydantic/pydantic) from 2.13.4 to 2.13.5.
- [Release notes](https://github.com/pydantic/pydantic/releases)
- [Changelog](https://github.com/pydantic/pydantic/blob/v2.13.5/HISTORY.md)
- [Commits](https://github.com/pydantic/pydantic/compare/v2.13.4...v2.13.5)

---
updated-dependencies:
- dependency-name: pydantic
  dependency-version: 2.13.5
  dependency-type: direct:development
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>


[…]
