### 2026-08-31 01d9fcb
Delete .github/workflows/joss.yml
