### 2026-08-31 c523289
Prepare draft for JOSS submission (#536)
* First draft of the paper

* Update draft

* fix width

* fixes

* fix nipoppy

* add citations

* add dates

* extend features section

* minor changes

* restructure the sections

* Fix spelling mistakes in JOSS draft (#560)

* Update URL for Adaptive Computing entry in bibliography (#561)

* Clarify target audience and benefits of pysqa (#562)

Expanded the statement of need to clarify the target audience for pysqa and its advantages over embedding scheduler-specific commands.

* Add reference for jinja2

* Delete .github/workflows/joss.yml
