+++ docs/paper/paper.bib
@@ -1,0 +1,158 @@
+@inproceedings{slurm,
+    author = {Morris A. Jette and Andy B. Yoo and Mark Grondona},
+    title = {SLURM: Simple {L}inux Utility for Resource Management},
+    booktitle = {In Lecture Notes in Computer Science: Proceedings of Job Scheduling Strategies for Parallel Processing ({JSSPP}) 2003},
+    year = {2002},
+    pages = {44--60},
+    publisher = {Springer-Verlag},
+    doi = {10.1007/10968987_3}
+}
+
+@article{pyiron,
+    title = {pyiron: An integrated development environment for computational materials science},
+    journal = {Computational Materials Science},
+    volume = {163},
+    pages = {24-36},
+    year = {2019},
+    issn = {0927-0256},
+    doi = {10.1016/j.commatsci.2018.07.043},
+    url = {https://www.sciencedirect.com/science/article/pii/S0927025618304786},
+    author = {Jan Janssen and Sudarsan Surendralal and Yury Lysogorskiy and Mira Todorova and Tilmann Hickel and Ralf Drautz and Jörg Neugebauer}
+}
+
+@article{myqueue, 
+    doi = {10.21105/joss.01844}, 
+    url = {https://doi.org/10.21105/joss.01844}, 
+    year = {2020}, 
+    publisher = {The Open Journal}, 
+    volume = {5}, 
+    number = {45}, 
+    pages = {1844}, 
+    author = {Jens Jørgen Mortensen and Morten Gjerding and Kristian Sommer Thygesen}, 
+    title = {MyQueue: Task and workflow scheduling system}, 
+    journal = {Journal of Open Source Software}
+}
+
+@article{jobflow,
+    doi = {10.21105/joss.05995},
+    url = {https://doi.org/10.21105/joss.05995},
+    year = {2024},
+    publisher = {The Open Journal},
+    volume = {9},
+    number = {93},
+    pages = {5995},
+    author = {Rosen, Andrew S. and Gallant, Max and George, Janine and Riebesell, Janosh and Sahasrabuddhe, Hrushikesh and Shen, Jimmy-Xuan and Wen, Mingjian and Evans, Matthew L. and Petretto, Guido and Waroquiers, David and Rignanese, Gian-Marco and Persson, Kristin A. and Jain, Anubhav and Ganose, Alex M.},
+    title = {Jobflow: Computational Workflows Made Simple},
+    journal = {Journal of Open Source Software}
+}
+
+@article{executorlib, 
+    doi = {10.21105/joss.07782},
+    url = {https://doi.org/10.21105/joss.07782},
+    year = {2025},
+    publisher = {The Open Journal},
+    volume = {10},
+    number = {108},
+    pages = {7782},
+    author = {Janssen, Jan and Taylor, Michael Gilbert and Yang, Ping and Neugebauer, Joerg and Perez, Danny},
+    title = {Executorlib - Up-scaling Python workflows for hierarchical heterogenous high-performance computing},
+    journal = {Journal of Open Source Software}
+}
+
+@inproceedings{psij,
+    author = {Hategan-Marandiuc, Mihael and Merzky, Andre and Collier, Nicholson and Maheshwari, Ketan and Ozik, Jonathan and Turilli, Matteo and Wilke, Andreas and Wozniak, Justin M. and Chard, Kyle and Foster, Ian and da Silva, Rafael Ferreira and Jha, Shantenu and Laney, Daniel},
+    booktitle = {2023 IEEE 19th International Conference on e-Science (e-Science)}, 
[…]
