+++ docs/paper/paper.md
@@ -1,0 +1,106 @@
+---
+title: 'pysqa - Python Simple Queuing System Adapter'
+tags:
+  - Python
+  - High Performance Computing
+  - Job Scheduler
+authors:
+  - name: Jan Janssen
+    orcid: 0000-0001-9948-7119
+    affiliation: 1
+  - name: Joerg Neugebauer
+    orcid: 0000-0002-7903-2472
+    affiliation: 1
+
+affiliations:
+ - name: Max Planck Institute for Sustainable Materials, Düsseldorf, Germany
+   index: 1
+date: 14 June 2026
+bibliography: paper.bib
+---
+
+# Summary
+High-performance computing (HPC) resources are commonly accessed through batch scheduling systems such as LSF [@lsf], Torque/PBS [@pbs], SGE [@sge],  SLURM [@slurm], Moab [@moab] and Flux [@flux]. These schedulers provide powerful command-line interfaces for job submission and resource management, but integrating scheduler interactions into Python applications often requires scheduler-specific scripts, shell wrappers, and custom parsing of scheduler outputs. As a result, scientific software frequently contains infrastructure code that is difficult to maintain and port across HPC environments.
+
+![Schematic overview of pysqa. The package provides a scheduler-independent interface for submitting, monitoring, and managing HPC jobs from Python or the command line. Reusable Jinja2-based templates are translated to the native syntax of supported queuing systems (e.g., SLURM, PBS, LSF, and SGE), while built-in support for remote connections, reservations, and job management simplifies the integration of HPC resources into scientific workflows.\label{fig:pysqa}](pysqa.png){width="100%"}
+
+pysqa (Python Simple Queue Adapter) is an open-source Python library that provides a lightweight, scheduler-independent interface for submitting, monitoring, and managing HPC jobs. The design of pysqa is based on reusable queue templates written in the Jinja2 templating language [@jinja2]. Users define scheduler-specific submission scripts once and subsequently reuse them across many computational tasks and workflows. This approach separates scheduler configuration from application logic while preserving transparency by generating native scheduler submission scripts. pysqa currently supports Flux, LSF, Moab, SGE, SLURM, and Torque queuing systems and includes optional support for remote job submission via SSH.
+
+Unlike workflow orchestration frameworks, pysqa focuses exclusively on the scheduler interaction layer. It enables computational scientists and research software engineers to integrate HPC resources into Python applications with minimal complexity while retaining full control over scheduler-specific configurations. A schematic overview of pysqa is illustrated in \autoref{fig:pysqa}.
+
+# Statement of Need
+Modern computational research increasingly relies on automated execution of simulations, machine learning workloads, and data-processing pipelines on shared HPC infrastructure. While scheduler command-line tools such as sbatch, squeue, and scancel provide direct access to HPC resources, scientific applications often require programmatic job submission and monitoring capabilities. Embedding scheduler-specific commands directly into software reduces portability and increases maintenance costs when users operate across multiple clusters. The primary audience for pysqa is developers of scientific software and research software engineers who require a lightweight, programmatic interface to HPC schedulers that can be used across different computing environments.
+
+# State of the field 
+Several software projects address related challenges. MyQueue [@myqueue] provides a higher-level task and workflow abstraction designed for scientific computing campaigns. PSI/J [@psij] offers a portable job execution API spanning multiple schedulers and execution backends. Jobflow-Remote focuses on remote execution of workflow graphs within the Jobflow ecosystem [@jobflow]. These tools provide broader workflow or interoperability capabilities, but they also introduce additional abstractions and infrastructure requirements.
+
[…]
