### 2026-08-31 #10961 @jan-janssen
@AnjaliSandip I merged the `paper` branch into `main`. The new release is:
https://github.com/pyiron/pysqa/releases/tag/pysqa-1.0.0

So the approved version number is 1.0.0 and the corresponding DOI: https://doi.org/10.5281/zenodo.22212817
