### 2026-09-02 #10961 @jan-janssen
@AnjaliSandip I merged the `paper` branch and deleted it, the paper is now available in the main branch. Still the editorial bot is looking for the `paper` branch.
