### 2026-09-03 #10961 @jan-janssen
In https://github.com/openjournals/joss-reviews/issues/10961#issuecomment-5511729639 the bot complained about an issue with the paper. I then updated the branch in https://github.com/openjournals/joss-reviews/issues/10961#issuecomment-5511886684 .

The error message was:
```
`failure': Couldn't check the bibtex because branch name is incorrect: paper (Theoj::Error)
```
