### 2026-09-03 #10961 @danielskatz
The editor should do this normally, but if you suspect some problem, I can try to jump in to help
