### 2026-09-03 #10961 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
