### 2026-09-03 #10961 @danielskatz
@AnjaliSandip - please confirm that this is ready for me to take over as track editor
